remove redundante code
