+++ WebApi/InWords.WebApi/InWords.WebApi/Controllers/v1/CardsGame/ScoreController.cs
@@ -18,12 +18,12 @@ namespace InWords.WebApi.Controllers.v1.CardsGame
     [Produces("application/json")]
     public class ScoreController : ControllerBase
     {
-        private readonly IGameScoreService gameScoreV2Service;
+        private readonly IGameScoreService gameScoreService;
         private readonly GameService gameService;
 
-        public ScoreController(GameScoreV2Service gameScoreV2Service, GameService gameService)
+        public ScoreController(GameScoreService gameScoreService, GameService gameService)
         {
-            this.gameScoreV2Service = gameScoreV2Service;
+            this.gameScoreService = gameScoreService;
             this.gameService = gameService;
         }
 
@@ -39,12 +39,12 @@ namespace InWords.WebApi.Controllers.v1.CardsGame
         {
             int authorizedId = User.GetUserId();
             // calculate score
-            LevelScore answer = gameScoreV2Service.GetLevelScore(levelResult);
+            LevelScore answer = gameScoreService.GetLevelScore(levelResult);
 
             // save score to user level
             try
             {
-                await gameScoreV2Service.UpdateUserScore(authorizedId, answer);
+                await gameScoreService.UpdateUserScore(authorizedId, answer);
             }
             catch (ArgumentNullException e)
             {
@@ -67,7 +67,7 @@ namespace InWords.WebApi.Controllers.v1.CardsGame
 
             try
             {
-                await gameScoreV2Service.PushLevelScoreList(authorizedId, levelScores);
+                await gameScoreService.PushLevelScoreList(authorizedId, levelScores);
             }
             catch (ArgumentNullException e)
             {
@@ -94,7 +94,7 @@ namespace InWords.WebApi.Controllers.v1.CardsGame
             if (answer == null) return NotFound();
 
             // set stars
-            answer = await gameScoreV2Service.GetGameStars(userId, answer);
+            answer = await gameScoreService.GetGameStars(userId, answer);
 
             return Ok(answer);
         }
