+++ WebApi/InWords.WebApi/InWords.WebApi/Services/GameService/GameScoreService.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using InWords.Data.Creations.GameBox;
@@ -13,40 +12,17 @@ namespace InWords.WebApi.Services.GameService
 {
     public class GameScoreService : IGameScoreService
     {
-        //TODO : return full game info
         public async Task<GameObject> GetGameStars(int userId, GameObject game)
         {
-            // find user game box to find all user levels
-            UserGameBox userGameBox = userGameBoxRepository
-                .GetWhere(usb => usb.UserId.Equals(userId) && usb.GameBoxId.Equals(game.GameId))
-                .SingleOrDefault();
-            // if no saves found return default game value
-            if (userGameBox == null) return game;
+            UserGameLevel[] levels = userGameLevelRepository.GetWhere(
+                g => game.LevelInfos.Any(i => i.LevelId.Equals(g.GameLevelId)) 
+                     && g.UserId.Equals(userId)).ToArray();
 
-            // load all saves
-            IEnumerable<UserGameLevel> userLevels =
-                userGameLevelRepository.GetWhere(ugl => ugl.UserGameBoxId.Equals(userGameBox.UserGameBoxId));
-            await SetLevelStars(game, userLevels);
-            return game;
-        }
-
-        private async Task SetLevelStars(GameObject game, IEnumerable<UserGameLevel> userLevels)
-        {
-            // merge by level number
-            foreach (UserGameLevel level in userLevels)
+            foreach (LevelInfo level in game.LevelInfos)
             {
-                LevelInfo userLevel = game.LevelInfos.Find(l => l.LevelId.Equals(level.GameLevelId));
-
-                if (userLevel == null)
-                {
-                    UserGameLevel nullGame = await userGameLevelRepository.FindById(level.UserGameLevelId);
-                    await userGameLevelRepository.Remove(nullGame);
-                }
-                else
-                {
-                    userLevel.PlayerStars = level.UserStars;
-                }
+                level.PlayerStars = levels.SingleOrDefault(l => l.GameLevelId.Equals(level.LevelId))?.UserStars ?? 0;
             }
+            return game;
         }
 
         public LevelScore GetLevelScore(LevelResult levelResult)
@@ -60,80 +36,88 @@ namespace InWords.WebApi.Services.GameService
 
         public async Task UpdateUserScore(int userId, LevelScore levelScore)
         {
-            UserGameBox userGameBox = await EnsureUserGameBox(userId, levelScore);
+            UserGameLevel level = userGameLevelRepository.GetWhere(ugl =>
+                ugl.UserId.Equals(ugl.UserId) && ugl.GameLevelId.Equals(levelScore.LevelId)).SingleOrDefault();
+            // if exist update if better
+            if (level != null)
+            {
+                if (level.UserStars > levelScore.Score) return;
+                level.UserStars = levelScore.Score;
+                await userGameLevelRepository.Update(level);
+            }
+            // add if not exist
+            else
+            {
+                await AddLevels(userId, levelScore);
+            }
 
-            await EnsureLevelScore(levelScore, userGameBox);
         }
 
         public async Task PushLevelScoreList(int userId, IEnumerable<LevelScore> levelScores)
         {
-            // TODO speedup (1)
-            // Thought like for 3 hours, and did not implement a rational search
-            foreach (LevelScore levelScore in levelScores) await UpdateUserScore(userId, levelScore);
+            // to prevent multiply enumerable
+            LevelScore[] levelScoresArray = levelScores.ToArray();
+            // find all that exist
+            UserGameLevel[] levelsExist = GetExistingLevels(userId, levelScoresArray);
+            // add any that not in existing in database
+            LevelScore[] levelScoresToAdd = GetScoresExceptExist(levelScoresArray, levelsExist);
+
+            await UpdateLevels(levelsExist, levelScoresArray);
+            await AddLevels(userId, levelScoresToAdd);
         }
 
-        private async Task<UserGameBox> EnsureUserGameBox(int userId, LevelScore levelScore)
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="levelsToUpdate">All levels that exist in database</param>
+        /// <param name="levelScores">All score that user send</param>
+        /// <returns></returns>
+        private async Task UpdateLevels(IEnumerable<UserGameLevel> levelsToUpdate, IEnumerable<LevelScore> levelScores)
         {
-            // Create user game stats
-            GameLevel gameLevel = await gameLevelRepository.FindById(levelScore.LevelId);
-
-            // if game level is not exits
-            if (gameLevel == null) throw new ArgumentNullException(nameof(gameLevel));
-
-            // find user game box that's contains user progress
-            UserGameBox userGameBox = userGameBoxRepository
-                                          .GetWhere(ugb =>
-                                              ugb.UserId.Equals(userId) && ugb.GameBoxId.Equals(gameLevel.GameBoxId))
-                                          .SingleOrDefault()
-                                      // create if not exists
-                                      ?? await userGameBoxRepository.Create(
-                                          new UserGameBox(userId, gameLevel.GameBoxId));
-
-            return userGameBox;
+            levelsToUpdate = from userGameLevel in levelsToUpdate
+                             join scores in levelScores on userGameLevel.GameLevelId equals scores.LevelId
+                             where scores.Score > userGameLevel.UserStars
+                             select new UserGameLevel()
+                             {
+                                 GameLevelId = userGameLevel.GameLevelId,
+                                 UserStars = scores.Score,
+                                 UserId = userGameLevel.UserId,
+                                 UserGameBoxId = userGameLevel.UserGameBoxId,
+                                 UserGameLevelId = userGameLevel.UserGameLevelId
+                             };
+            await userGameLevelRepository.Update(levelsToUpdate.ToArray());
         }
 
-        private async Task EnsureLevelScore(LevelScore levelScore, UserGameBox userGameBox)
+        private async Task AddLevels(int userId, params LevelScore[] levels)
         {
-            // find user game level
-            UserGameLevel userGameLevel = userGameLevelRepository
-                .GetWhere(g =>
-                    g.GameLevelId.Equals(levelScore.LevelId) && g.UserGameBoxId.Equals(userGameBox.UserGameBoxId))
-                .SingleOrDefault();
-
-            // if user level score information not found then create save
-            if (userGameLevel == null)
+            UserGameLevel[] userGameLevels = levels.Select(levelScore => new UserGameLevel()
             {
-                userGameLevel = new UserGameLevel(userGameBox.UserGameBoxId, levelScore.LevelId, levelScore.Score);
-                await userGameLevelRepository.Create(userGameLevel);
-            }
-
-            // if level score information found 
-            else
-            {
-                // if score less or equals return current score do nothing
-                if (userGameLevel.UserStars >= levelScore.Score) return;
-
-                // else update score
-                userGameLevel.UserStars = levelScore.Score;
-                await userGameLevelRepository.Update(userGameLevel);
-            }
+                UserId = userId,
+                UserStars = levelScore.Score,
+                GameLevelId = levelScore.LevelId
+            }).ToArray();
+            await userGameLevelRepository.Create(userGameLevels);
         }
 
-        #region Ctor
+        private UserGameLevel[] GetExistingLevels(int userId, LevelScore[] levelScores)
+        {
+            return userGameLevelRepository.GetWhere(ugl => ugl.UserId.Equals(userId)
+                                                     && levelScores.Any(ls => ls.LevelId.Equals(ugl.GameLevelId))).ToArray();
+        }
 
-        private readonly UserGameBoxRepository userGameBoxRepository;
+        private static LevelScore[] GetScoresExceptExist(IEnumerable<LevelScore> levelScoresArray, IEnumerable<UserGameLevel> levelsExist)
+        {
+            return (from ls in levelScoresArray
+                    where !levelsExist.Any(ltu => ltu.GameLevelId.Equals(ls.LevelId))
+                    select ls).ToArray();
+        }
+
+        #region ctor
+
         private readonly UserGameLevelRepository userGameLevelRepository;
-        private readonly GameLevelRepository gameLevelRepository;
-
-        public GameScoreService(
-            UserGameBoxRepository userGameBoxRepository,
-            UserGameLevelRepository userGameLevelRepository,
-            GameLevelRepository gameLevelRepository
-        )
+        public GameScoreService(UserGameLevelRepository userGameLevelRepository)
         {
-            this.userGameBoxRepository = userGameBoxRepository;
             this.userGameLevelRepository = userGameLevelRepository;
-            this.gameLevelRepository = gameLevelRepository;
         }
 
         #endregion
